server/requirements.txt
sqlalchemy==2.0
pydantic==1.10
fastapi==0.95
mysqlclient==2.1
passlib==1.7
python-jose[cryptography]==3.3